WL#6043 - Optimizations for InnoDB's FTS
The optimization where tables rows are not accessed when only
document ID and ranking score was requested, did not apply when
there was no ORDER BY clause.  This is fixed by this patch.

The reason for the issue was that when there was no ORDER BY
clause, a flag FT_SORTED was set to signal to the storage engine
that it should do the sorting.  This caused the optimizer to
believe that the MATCH expression in the WHERE clause and the 
SELECT list were not equivalent.  The fix is to not consider
the FT_SORTED flag when considering equivalence.

Activating this optimization revealed two bugs than needed to
be fixed:
1. Item_func_docid needs to set const_item_cache to false to
   prevent the optimizer from considering this function to be
   constant.
2. The optimization has not been implemented to handle queries
   that contain a GROUP BY clause.