BUG#13971348 - FAILURE WHILE CONFIGURING REPLICATION CAUSES THE SERVER TO SHUTDOWN
If something went wrong while setting up replication's objects (i.e. slave 
objects), the server was brought off-line. For some specific errors, one
could provide --skip-slave-start to try to avoid shutting down the server.

This was a drastic measure as the replication is just one of the services 
provided by the server.

With this patch, we just print out some error messages on the problem and
let the server continue its execution.