WL#5223: Binary Log Group Commit
Patch to remove the last traces of the prepare_commit_mutex
from InnoDB code.

This patch adds a function thd_binlog_pos that gives the file
and position for the transaction last written to the binary
log and uses that in innobase_commit.

The old functions that were declared in InnoDB-specific files
are removed.