Bug#13989037 - AVOID CONTEXT SWITCH DURING TRX_COMMIT() REMOVE THE VIEW EARLY
Currently when committing a transaction we release and reacquire the 
trx_sys_t::mutex when we remove the transaction's read view from the global 
view list. This will cause unnecessary context switching. It is an O(1) 
operation on a linked list. Better to do it where we have have already 
acquired the trx_sys_t::mutex. 

Approved by Jimmy, rb://1040