Bug#13993661 - GET RID OF STATIC COUNTER IN TRX_START_LOW(), USE THE TRX_T::ID AS THE COUNTER
In trx_start_low() we use a static counter to control the number of 
gettimeofday() calls we make to set the trx_t::start_time. The trx_t::id is 
also a monotonically increasing counter and is local to the thread. Use that 
instead. This should reduce potential cache coherency issues.
      
Approved by Inaam over IM.