Bug#14004341 - REDUCE MEMCPY() OVERHEAD IN ROW_SEARCH_FOR_MYSQL()
InnoDB uses a read ahead cache for rows that can potentially be accessed in a 
batch. This is to avoid the overhead of mini-transactions. Currently, all 
records are copied to this fetch cache and when we exit from the function 
we copy the first record from this queue to the MySQL buffer. Avoid this copy 
to the fetch cache queue and back on normal exit from the function by doing the 
changes directly in the MySQL recod buffer for the first record. 

This eliminates two memcpy() calls but is only done for non-ICP code paths.

rb://1046 Approved by Jimmy.