ha_ndbcluster::read_first_row() was once implemented as part of the SPJ project as we then had plans for pushing queries with 'system' tables as root operation. (JT_SYSTEM is the only use case for ::read_first_row())
It was later decided that we should not push such queries, but
ha_ndbcluster::read_first_row() has remained in the code since then.

There is a default implementation of this method (handler::read_first_row()) which
is almost identical, except that it does not call ha_rnd_end() / ha_index_end()
before returning. These calls was removed in our ha_ndbcluster::read_first_row()
as pushed joins required the resultset to stay 'open' due to the child operations.

However:
 - We do not push queries with 'system' tables so such a change
   is not required.
 - As part of fixing bug#13901890 'SQL NODE CRASHES DURING INSERT SELECT'
   we identified part of the root cause to be an overridden
   ::index_read_idx_map() which also omitted calling rnd_end()/index_end()
   before returning.

This fix removes this obsolete, and potential evil code.