Bug#14034087 - FIX FALSE CACHELINE SHARING IN 5.6
1. Reduce false sharing

2. Reduce the overhead of cache coherency

3. Remove redundant code

4. Remove bogus stats printfs that were only enabled when UNIV_DEBUG was set

Add a class ib_counter_t that splits the counter into N slots. Currently
hard coded to 64 slots. It pads the slots on 64 bytes, again currently hard
coded. We may in the future make these configurable once there is a platform
independent and reliable way to get the necessary information.

There is an implementation that uses the Linux function cpu_getsched() instead
of the OS thread id. It is currently not used.

Group all the stats variables in srv_stats_t.

rb://1028 Approved by Inaam Rana.