Some of the SPJ-MTR tests are written to detect failures (previously) caused by incorrect handling when the result was retrieved in multiple batches. Thus, these tests depends in a specific 'batch size' being used.
As the default batchsize was increased from 64 -> 256 rows in cluster-7.2
we believe these tests to now be broken.

This fix introduce the debug flags 'max_4rows_in_spj_batches' and
'max_64rows_in_spj_batches' which will enforce the usage of the
batchsize as indicated by their names.

This fix is also a preparation for making a fix + testcase for:
Bug# 14010406 LARGE PUSHED JOIN HIT ASSERT IN SPJ BLOCK