part of WL#6189 Turn InnoDB persistent statistics ON by default
Synchronize the background stats thread with DROP TABLE. The latter should
wait for the background stats thread to stop working on the table and close
it before proceeding. Otherwise the DROP TABLE code may see that the table
is opened and add it to the background drop queue. This is not acceptable
because, an immediate CREATE after DROP may fail with a "Table already
exists" error. For example:

CREATE TABLE
INSERT 1 row
INSERT 1 row --> will trigger background stats gathering, which will open
                 the table
DROP TABLE --> will see the table is opened, will emit a notice in the log
               that MySQL is trying to drop a table that has opened handles
	       to it and will add it to the background drop queue
CREATE TABLE --> fails because the table is not yet removed by the background
                 drop code