WL#5223: Binary Log Group Commit
This patch add a variable where the user can choose
between running commits either in parallel or ordered
in sequence.
      
When ordered, the transactions are committed in the same
order as the transactions are written to the binary log.
When committed in paralle, each thread execute an engine
commit independent of the other sessions.