Bug#13970015: ASSERT `MIN_ENDP || MAX_ENDP' FAILED IN               HANDLER::MULTI_RANGE_READ_INFO_CONST
When key_or() finds that two SEL_ARG trees form a full range
(no upper or lower bound in the range; is always true), NULL 
is returned. In most cases the range optimizer then forgets 
about these trees, but when the trees belong to two different
SEL_TREEs used in an index merge the always true range is only
removed from one of the SEL_TREEs. 

Index merge is a disjunct operation, meaning that the rows
read from the different indexes shall be ORed. If the
predicate for one of the indexes is always true then index 
merge range access will effectively fetch all rows from the
table. Range access in this case is not a good idea.