Fix for Bug#14010406 LARGE PUSHED JOIN HIT ASSERT IN SPJ BLOCK
Buffered rows (caused by PARENT refs) was prematurely
released when there was no TN_ACTIVE treeNodes childs.

However, in bushy scans, a scan branch can be 'repeated'
even if its previous execution was not 'ACTIVE'

Thus, the logic for when buffered row should be released should be
less eager to release rows. Now we keep any buffered rows until
we prepare the treeNodes to retrieve new rows for the specific treeNode.

This also made it possible to simplify this logic such that the
recursive method Dbspj::cleanupChildBranch() became obsolete.

As part of debugging / fixing this problem, SPJ-DEBUG trace was
enhanced such that execution of SPJ signals are related
to 'node_no' of treeNode and ident of the request it belongs to.