Bug#14006907 FOREIGN KEY problems after DROP INDEX
dict_index_t: Add the to_be_dropped flag to incidate when an index
is about to be dropped.

dict_foreign_find_index(): Assert that the caller is holding the
dict_sys->mutex. Skip indexes that are flagged as to-be-dropped.

dict_foreign_replace_index(): Assert that the index is to-be-dropped.

ha_innobase::prepare_inplace_alter_table(): Lock the data dictionary
while checking if any foreign key constraints depend on the indexes
that are to be dropped. Flag them as such in the data dictionary cache.

ha_innobase::inplace_alter_table(): Add a debug sync point for the
DEBUG_SYNC test.

rollback_inplace_alter_table(): Clear the index->to_be_dropped flag in
the data dictionary cache.

rb:1049 approved by Jimmy Yang