Bug#14057307 - CRASH DUE TO FIX FOR BUG#14034087
The event that coordinates between the lock wait monitor thread and other 
threads needs to be anchored in lock_sys_t. The current problem is that at 
startup, the monitor thread may not have started up. To ensure that the event 
always exists we create it at the same time as the lock data structures are 
created at startup.

rb://1066 Approved by Jimmy Yang.