Change RENAME TABLE where it renames the table stats to use UPDATE insead of delete and save because at the time dict_stats_save() is called the table be closed and in-memory stats cleared.
UPDATE was not used before because of the foreign key which existed
between the stats tables. The foreign key is now removed.

As a side effect, the result in the exceptional case where rows exist
for the new table name is changed. Instead of overwriting existing rows
(which was the case with delete and save()) now the existing rows are
left and RENAME TABLE emits a warning that it failed to rename the table
stats.