part of WL#6189 Turn InnoDB persistent statistics ON by default
Temporarily disable background stats gathering until locking issues are
resolved. The code is sound as it is and can be pushed in that state.

The disadvantage is that the query which triggers the 10% thresold will
be blocked until the new stats are gathered. Also, since there is no queue,
all events that trigger the 10% thresold MIN_RECALC_INTERVAL seconds after
stats were updated will be ignored insead of queued for later processing.