BUG#13463397 - 63562: UNKNOWN DATABASE INFORMATION_SCHEMA
Regression Test added to demonstrate that main complaint 
mentioned in the bug report CREATE PROCEDURE information_schema.is() 
giving "Unknown database 'information_schema' is not reproducible 
in 5.6. It currently gives ER_DBACCESS_DENIED_ERROR.

Closing this bug as it is not reproducible in 5.6 and attaching 
a small note to address the other complaints mentioned in 
the bug report. 

NOTE:
This is a small testcase written to demonstrate the inconsistencies 
of error messages for I_S and P_S databases. This is the current 
behaviour of I_S and P_S in 5.6.

--error ER_DBACCESS_DENIED_ERROR
CREATE PROCEDURE information_schema.is() BEGIN END;
--error ER_DBACCESS_DENIED_ERROR
CREATE PROCEDURE performance_schema.ps() BEGIN END;

--error ER_DBACCESS_DENIED_ERROR
CREATE TABLE information_schema.is(a INT);
--error ER_TABLEACCESS_DENIED_ERROR
CREATE TABLE performance_schema.ps(a INT);

--error ER_DBACCESS_DENIED_ERROR
ALTER TABLE information_schema.collations ADD COLUMN c INT;
--error ER_DBACCESS_DENIED_ERROR
ALTER TABLE performance_schema.setup_consumers ADD COLUMN c INT; 

Ideally Both of I_S and P_S databases should give same errors 
for the same operation but this is currently NOT the case.

These inconsistencies will be handled later when the runtime 
team starts working on New Data Dictionary Project as it is 
pretty hard to fix these inconsistencies at this point of time 
in the code.