Bug#13966809 CRASH IN COPY_FUNCS WHEN GROUPING BY OUTER QUERY BLOB FIELD IN SUBQUERY
Problem: Grouping by an outer BLOB field in a subquery causes the
server to crash.

The outer field is referenced through an Item_outer_ref that points to
an Item_ref that points to an Item_copy_string. In create_tmp_table(),
an array is created to store pointers to function items that are
replaced with fields. The size of this array is determined by a count
of functions, passed as parameter to create_tmp_table(). When
create_tmp_field() is called for the outer field, it calls
create_tmp_field_from_item(), where the field is mistaken for a
function, and a pointer to the Item_outer_ref is stored in the array
of function items. This results in overflowing the allocated storage
space since one more item than intended is added.

The mistake is made because create_tmp_field_from_item() checks
item->is_result_field() to determine if the item is a function. Since
the Item_outer_ref, like functions, returns true for this test, it is
assumed to be a function, regardless of what item type it is a
reference to. This bug is triggered only when the item referred to by
the Item_ref is neither a field and nor a function.

In copy_funcs(), a terminating null pointer is expected to end the
array of function items. Because of the extra entry in the array, this
null pointer is outside the allocated storage space for the array and
the memory location may be overwritten with other information, causing
copy_funcs() to dereference a wild pointer.

Fix: Check item->real_item()->is_result_field() to determine if the
item should be added to the array of function items.