Bug#13423851: CRASH IN ITEM_FIELD::ITEM_FIELD, WITH QUERY USING A VIEW
Queries with an ORDER BY clause are frequently executed
using a temporary table, where some of the expressions from
the select list are stored. The function create_tmp_table()
chooses all select list expressions that are not constant
during evaluation of the query (and some additional ones.)
For column references that are resolved to a view which will
be "merged", the Item_direct_view_ref object is asked, which
checks that the expression does not reference any
tables. The bug made it fail to include RAND_TABLE_BIT into
the set of tables if the expression is a non-deterministic
function.

When execution begins, the data in the temporary table are
needed. We now try to yet again infer the column's
const-ness. But this time round the reference has been
flattened (since the view was merged) and we are talking to
the Item_func directly. And it knows that it's always
non-constant, so obviously it thinks that create_tmp_table()
must have set something up. But since it didn't, all we find
is a NULL pointer.

This fix implements a means to make sure RAND_TABLE_BIT is
always set in used_tables() by creating a new interface
get_initial_pseudo_tables(), the result of which must always
be picked up whenever we cache the set of used tables. Hence
all assignments of Item_func::used_tables_cache= 0 have been
replaced by Item_func::used_tables_cache=
get_initial_pseudo_tables(). This enables us to remove some
other, more error-prone ways of setting RAND_TABLE_BIT.

We make some changes to an optimizer_trace test:

A non-deterministic stored procedure is used in a WHERE
equality condition inside a subquery. The subquery is
rewritten into a semijoin. Because the optimizer now
correctly recognizes the stored procedure as
non-deterministic, it does not consider using an index
access on the outer table. This causes DuplicateWeedout to
be used rather than the previously used FirstMatch. Because
the strategy involves access to a temporary table, we need
to make more room in the optimizer trace depth.