Bug #14075825 LCP Watchdog : Fragment scan check
The Local Checkpoint (LCP) mechanism allows Ndb to trim its redo and
undo logs.  This keeps node and system recovery time within bounds, 
and maintains free redo and undo space to support logging DML 
operations. 
 
A watchdog mechanism is added to supervise fragment scans occurring
as part of a local checkpoint.  This is intended to guard against any
scan related OS level IO errors or bugs causing issues with LCP and 
endangering write service and recovery times.

Each node independently monitors the progress of local fragment 
scans occurring as part of an LCP.  If no progress is made for 
20 seconds, warning logs will be generated every 10 seconds.  
If no progress is made for one minute then the fragment scan is 
considered to be hung, and the node will be restarted to enable 
LCP to continue. 

Any occurrence of warning logs or other action taken by this 
watchdog should be reported as a bug.

A new test scenario is added to testNodeRestart.