Bug#14117641 (possible fix) and a WL#6255 optimization
Allow the non-online InnoDB in-place DDL operations with LOCK=SHARED.

prepare_inplace_alter_table_dict(): When creating indexes and
an online operation has not been requested, acquire S-lock on the table.
Previously, we acquired an X-lock, and only when the table was being rebuilt.
Failure to lock the table might explain Bug#14117641 if there was a
foreign key CASCADE or SET NULL constraint referencing the table.