WL#6255 preparation: Add assertions to purge code.
row_purge_remove_clust_if_poss_low(), row_purge_upd_exist_or_extern_func():
Assert that dict_operation_lock is being S-latched. Narrow the scope
of some variable declarations.

In online table rebuild, there is a potential problem in BLOB retrieval.
The planned solution is that we will have a lookup table of trx_id_t of
rolled-back transactions. Any rows inserted or updated by such transactions
will be ignored. A delete-mark operation will be buffered as a delete+purge
in the online log, with the new PRIMARY KEY and the DB_TRX_ID of the record,
and applied by looking up the fields in the rebuilt table.
This (and the fact that the log apply will hold an S-latch or X-latch on
the clustered index tree of the old table) should guarantee that the
log apply will never dereference freed BLOBs in the old table.