WL#6082 Improve the Disk-Sweep Multi-Range Read cost model
Patch 2: Fix bug in how IO and CPU cost are computed for multi-pass scans. 
      
When computing the cost for each scan, the record estimate for the
total number of records is used. This causes the IO cost and CPU cost
to become too high. Change this to use the number of records that
will be read per scan.