part of WL#6189 Turn InnoDB persistent statistics ON by default
Disable rpl_row_binlog_max_cache_size mtr test because it hangs if
estimated number of rows is 3 instead of 4. With persistent stats the
estimated number is 3 (which is correct). We could change the test to
use transient stats (old method, which estimates the number of rows
incorrectly to 4) and the test will pass, but it is better to have
this test reviewed by either Optimizer or Replication team and adjusted
so that it does not hang, depending on the estimated number of rows.

The following bug has been opened to track this issue:
Bug#14126780 RPL_ROW_BINLOG_MAX_CACHE_SIZE HANGS IF ESTIMATED NUMBER OF
ROWS IS 3 INSTEADOF 4