Bug#14037167: OUTPUT ON FLUSHED PAGES IS NOT ACCURATE WITH MULTIPLE BUFFER POOL INSTANCES
rb://1070
approved by: Sunny Bains

When doing flushing buf_flush_list() returns ULINT_UNDEFINED if it
fails for any buffer pool instance because a batch of same type is
already running. On success it returns the number of pages
successfully queued for flushing. When using multiple buffer pool
instances it is possible that a flush batch was queued for some
instances in which case the accounting for flushing of these pages
gets wrong in the callers because ULINT_UNDEFINED is taken to mean
that no pages have been flushed.

This patch fixes that making return type of buf_flush_list() bool
and returning the number of pages as a separate parameter.