Bug #13485416: CRASH WITH SMALL MAX_SORT_LENGTH AND GROUP BY AND LIMIT
Sort keys for e.g. filesort are created by
Field::sort_string(uchar*,uint), where the uint argument is
ignored. But the three callers of the function obviously
assume the opposite: when max_sort_length is less than the
pack size of the value to be sorted by, there will always be
a buffer overrun.

The fix is to:

- Rename Field::sort_string() to Field::make_sort_key(),
  because the latter describes what the function does.

- Make sure we always respect the length argument for all
  implementations of the interface Field_*::make_sort_key().

- Test all implementations of make_sort_key with unit
  tests. Some implementations need to be rewritten
  (i.e. code in platform dependent #ifdef's) in order to be
  completely tested.