WL#6082 Improve the Disk-Sweep Multi-Range Read cost model
Patch 3: Memory cost too large

The cost of using memory for the sort buffer dominates the total cost
for many queries. Since we do not consider cost of using buffer space
in other places in the server we should not include it here either.

This patch removes inclusion of memory cost for the sort buffer
in the cost for doing ds-mrr scans.

This change causes a few changes in plans for existing tests.