WL#6255 MVCC bug fix: Ensure that older transactions are denied access to a table that has been rebuilt, whether or not it happened in-place (ALGORITHM=INPLACE) or in the old way (ALGORITHM=COPY).
When a table is created, we record the trx_id of the creating data
dictionary transaction in each index of the table in the data dictionary
cache. In this way, transactions that were started before the table
creation (or copying) can be denied access to the new table.

To be exact, there are two transactions involved in DDL. One is the
user transaction associated with the MySQL client connection. For
InnoDB data dictionary operations, we create a data dictionary transaction.
Theoretically, other transactions could start between the prebuilt->trx
and the dictionary trx. However, the table cannot be modified during this
time frame, thanks to a meta-data lock (MDL). In CREATE TABLE, the table
does not exist before InnoDB returns, and in ALTER TABLE, other transactions
are blocked from accessing the table while the table copy is being set up.

dict_create_index_step(): Initialize index->trx_id, so that older
transactions will be denied access to a newly created table.

ha_innobase::commit_inplace_alter_table(): Remove the adjustment of
index->trx_id for ALGORITHM=INPLACE.