Bug#14135054 ASSERT IN BUF_PAGE_IN_FILE(BPAGE) DURING DROP DATABASE
Needs the buffer pool mutex to be held for the duration of the fetch count
of pages matching some space id. When we relocate a page, first we change the
state to remove from hash and then remove the page from the flush list.
We release the page mutex after changing the state, therefore holding it
is not sufficient. We must hold the buffer pool mutex while scanning the
flush list. Note: This is debug code.
					        
Approved by Marko over IM