WL#6255 latching order bug fix.
row_ins_sec_index_entry_low(): Move the log_free_check() to the caller,
row_ins_sec_index_entry(). In online rebuild log apply, we are holding
an X-latch on the clustered index tree of the old table most of the time,
and must do the log_free_check() when not holding the latch.