Bug#14083116 (Bad error handling in LQH corrupts transid hash)
Early exits when processing LQHKEYREQ signals could result in 
corruption of the LQH transid hash, due to the wrong error handling
path being used.

This patch fixes the early exits, and adds assertions to the transid
hash code to help catch such problems quicker in future.

Testcase to be added on merge to cluster-7.2.