Bug#14145138 work-around for WL#5545 ALTER_COLUMN_NAME (rename column).
When a table contains or is referenced by FOREIGN KEY constraints, and
the table is rebuilt (ADD PRIMARY KEY or otherwise) while renaming
columns, the columns were not renamed in the FOREIGN KEY constraints.
This would cause an implicit DROP TABLE (Bug#14145138), because the
dict_load_foreigns() call in row_merge_rename_tables() would fail.

This fix does not address Bug#14145138, it just prevents one scenario
leading to it.

innobase_rename_column(): Add the parameter new_clustered, so that
column names in FOREIGN KEY constraints can be renamed when rebuilding
a table. Previously, we only renamed columns if the table was not being
rebuilt.

innobase_rename_columns(): New function, to invoke
innobase_rename_column() on all renamed columns. Invoke this also when
rebuilding a table (new_clustered==true).