Bug#14046403: SEGFAULT IN RESOLVE_SUBQUERY AT SQL/SQL_RESOLVER.CC               ON 2ND EXEC OF PREPARED STMT
In Item_in_optimizer::transform(), args[0] is assigned to 
Item_in_subselect->left_expr. In this bug, args[0] contained
a replacement item that was rolled back for each PS execution.
Because this rollback was not done for
Item_in_subselect->left_expr, it pointed to a deleted item on
second execution.

The fix is to rollback Item_in_subselect->left_expr as well. This
works fine because the transformation is done on each execution.