WL#6082 Improve the Disk-Sweep Multi-Range Read cost model
Patch 4: CPU cost is too high

The cost estimate for sorting the keys in the sort buffer for a large
scan is too high. It can be higher than the IO cost estimate for
reading the data from the database and in can be several times higher
than the CPU cost estimate for the same scan using standard MRR. In
addition it does not include the normal row processing cost that is
done in a standard MRR scan. So for scan that read just a few records,
the CPU cost estimate for DS-MRR can actually be lower than the CPU
estimate for standard MRR.

This patch fixes the following issues:
      
-a bug in the formula for calculating the CPU estimate for the sort
 based on quick sort.

-reduces the cost for doing the sort operation by replacing the use of
 ROWID_COMPARE_COST constant (value 0.1) with a new constant
 ROWID_COMPARE_SORT_COST that has the value =0.01.
      
-add the cost for doing row processing in the same way as it is
 calculated for stand MRR scans.
      
This patch changes execution plan for several test cases that are run
with mrr_cost_based off. The cause for these changes is that the cost
estimate for doing DS-MRR has changed (either increased or decreased).